Remove references to home app
Removed home from INSTALLED_APPS, removed its views from urls.py, and
removed django-ordered-model from INSTALLED_APPS and pip requirements.
Adrienne is working on replacing the home app with a hard-coded home
page from a template html file, so this app with homeContent objects is
deprecated. If we were also version-controlling the sqlite db, I would need
to make changes to that, but instead this requires a removal/rebuilding
of the database.