Improve SourceTrait and related models
Previously, SourceTrait and SourceEncodedValue models were completely
dependent on the schema from snuffles. This made the site code very
inflexible with respect to snuffles schema changes. In the interest of
separating the snuffles code and the phenotype_inventory code, I made
significant changes to the SourceTrait and SourceEncodedValue models. This
required making significant changes to templates, views, ModelAdmin
objects, and the populate_source_traits management command. I also changed
the name of this command from populate_trait_models to
populate_source_traits to more accurately reflect the purpose of this
command.