Add link over home page glyphicon
It was annoying me to not be able to click on the glyphicon, which changes
color on hover, so I added a link. It goes nowhere right now because we
still need a landing page for the study source traits, but that can be fixed
once we've made that page. We will still need to add these links on the
harmonized traits and harmonization request glyphicons when they are active.