Style for SourceTraitTable
Switch to using a custom template for the SourceTraitTable class. Create that
custom template (copy from the bootstrap.html included with django-tables2).
Add custom CSS to style.css for displaying glyphicons as the sorting arrows
for sorting SourceTraitTables. Move verbose_names to the SourceTraitTable
class (previously in the model fields). Turn off sorting for description field.