Fix default site name using data migration to sites app
The default behavior of the sites app + flatpages app is to create a
site with SITE_ID 1 that has domain name example.com. Without alteration,
we would have to change this in the admin interface every time we
set up the site code. This commit adds a data migration for the sites
app to change this default value to topmedphenotypes.org. It will be applied
the very first time manage.py migrate is run.

To implement this change, a core app was added and the migrations directories
were removed from the .gitignore file so that migrations are back under
version control.