Link SourceTrait model instances to a Study model with a foreign key
Added a foreign key linking the SourceTrait model instances to a Study
model. I had to change the SourceTrait detail_iter method because study_name
was no longer a field (had to get the name from the linked Study instead).
Changes were necessary in the admin.py script for the SourceTrait model for
the same reasons.