Add search feature for source traits
The search feature is working, but not ideal. I added crispy forms to the
installed apps and added a forms.py file to track the forms. I also modified
the view and template to actually perform a search and show the results.
However, the django-tables2 rendering of the results does not work when you
try to resort by a different column; it just reloads the page and erases
the results. I would also like to make the search a method (of Trait?
in the view?) so that it can be used for both souce traits and harmonized
traits in the future.