Implement SECRET_KEY environmental variable local setting
To further improve on the settings module, add a setting in settings.local to
get the SECRET_KEY from the locally stored environmental variable
DJANGO_SECRET_KEY. This takes the secret key setting out of version control
but still obtains and stores it in a consistent way.

Also add a script settings/generate_key.py to generate and print a new
random crypto key that users can store in their own environmental variables.