Implement tested & working staging settings
To implement the settings module, these changes introduce two separate
wsgi.py files, one for the staging apache server and one for the
production apache server. The settings in
phenotype_inventory.settings.staging are now tested and working on the
staging system. A compensatory change in the apache virtual host configuration
file was also made to point to staging_wsgi.py rather than wsgi.py, but this
change is untracked.

A system for storing secret settings on the staging and
production servers is also implemented in the get_secrets() function and
.secrets.json file. This secrets file is untracked and is to remain on the
server for which it applies.

Production settings and wsgi.py files are modified with this change, but
these settings are not tested. Production settings are basically placeholders
at this point.