Implement model_mommy for generating test model values
This set of changes introduces model_mommy to make test instances of Studies and
SourceTraits for testing. The interface is very simple and straightforward and
it works pretty well... except that about half of the time mommy.make throws an
error related to the NOT NULL constraint on web_date_added. I can't figure out
why this error is happening or how to fix it. I'm committing this set of changes
so the work isn't lost, but next I'll remove the model_mommy usage.