Change type checks to isinstance() tests in populate_source_traits fixers
Previous implementation of _fix_bytearray and _fix_timezone did checks to
look for type(thing) == desired_type, but this approach doesn't allow for
inheritance. This was discovered when a datetime object didn't show up as
a datetime, because it had come from Faker.datetime(). The new
implementation works well with inheritance.