Move STUDIES choices for search form to __init__ method
Previously, the SourceTraitCrispySearchForm study field generated choices as an
attribute of the class. This meant that the STUDIES choice list was only
generated once at django startup. This meant that the STUDIES could not change
after the webserver started running. This means that adding studies to the db
would require a server restart to update the study choices and that tests didn't
run properly (the Form object pulled the study choices from the real db rather
than the test db set up for each test method).

By moving the STUDIES choice generation to the overloaded __init__ method, this
problem is circumvented. Now the STUDIES choices list is generated each time a
form object is instantiated. This may end up having performance tradeoffs later
on, so caching of the form may be required at some later point.