Update source_trait_search function for empty input
Previously, the source_trait_search view function used a test of whether
request.method == 'get', but this condition is always true since we're not using
the POST method for this form. As a result, the form always showed an error for
no search text being entered in.

To fix this, the view now tests whether the request has any form input for the
search text. This change makes it so that you can view the search page without
error message, as long as no text has been entered. The error message only shows
up if you hit the Submit button without entering search text.

This also allows proper testing of the view function. And it makes the tests of
the form itself more applicable to the proper function of the view.