Update snuffles connection settings
Use .cnf files to connect to the snuffles phenotype database. Modified local
and production settings files so that CNF_PATH is coded into each. In
production it will be stored in the secrets json file, and in local it
will just look at the user's home directory for the .mysql-topmed.cnf file.