Implement tests that catch errors in two overlooked functions.
SourceEncodedValue.get_source_trait_name() and
SourceEncodedValue.get_source_trait_study() were not previously covered in tests
and this resulted in an undetected error. This error only shows up in the admin
interface and prevents the SourceEncodedValues admin page from displaying. One of
the new tests catches this error, and the other will catch errors in another
related function that was also overlooked.