Fix test of search form with empty search string
Previously, the test that was supposed to test the search form taking
an empty search string instead got no submitted form data. This resulted
in a lack of testing coverage over part of the search view. I changed
the test function to get an empty string text input instead and this
fixed the lack of testing coverage.