Use n_studies & n_traits arguments to limit records retrieved from source db
Use n_studies in _populate_studies with a LIMIT statement to limit the number
of studies imported from the source db. Use n_traits in _populate_source_traits
to limit the number of traits per study imported from the source db. If n_traits
is not set, restrict retrieved traits to those that match the previously imported
study primary keys. Change _populate_source_encoded_values to only retrieve values
for SourceTraits that are already in the django site db.