Add 'New Search' button on search page, replacing non-functional Reset
The Reset button in the crispy form for the source trait search form
was nonfunctional. Changing the button class made it functional, but the
function - clearing test from the search text input box - was pretty
unnecessary. Adding a button to the form that redirects to the empty
search form didn't work (the class is created before the urlconf is
loaded, and even reverse_lazy didn't work for this scenario). The
New Search button was added as a button in a new form on the search page
that is empty.