Improve SourceTrait detail view and template
Clean up the formatting of the SourceTrait detail view by
adding some nice panel and well classes to the template.
Remove the hacky detail_iter function and instead hardcode
the fields you want to show into the template. Remove the
unused field_iter function as well, along with tests of the
functions that were removed.