Fix bug where home, about, and contact URLs didn't have to be terminal
For the home, about, and contact URLs, I'd forgotten to put a $ at the
end of each of the URLs. This meant that if you added a bunch of stuff
after the given URL, it would still go to the home, about, or contact
page. E.g. /home/a/bunch/of/other/stuff/there/isn't/an/urlpattern/for/
would still send you to the home view. This change fixes this.