Add StudyTable class and change table attributes for SourceTraitTable
Add a class, StudyTable, to display a list of studies as a table.
Use some style and class attributes to set the study table to look
different than the SourceTraitTable where necessary. Add the
table-condensed class to the SourceTraitTable Class, to reduce the
amount of padding around cells and make the trait tables look cleaner.