Remove class assignment from table in bootstrap table template
The way the bootstrap table template was written, it was impossible
to change or add to the class of one of the tables. All of the
bootstrap table classes I had carefully chosen weren't showing up
properly. So, remove the class="" assignment from the bootstrap
table template and require the manual setting of table classes in
the attrs = {} Table.Meta field.