Implement URL namespace for trait_browser app
Add namespace='trait_browser' argument to the include statement for
the trait_browser app in the root urls.py file. This makes it so that
reverse(urlname) and {% url urlname %} can use shorter and more
specific urlnames, and there can be much shorter urlnames in the
trait_browser.urls module.