Reorganize and rename URLs and views
Reorganize URL structure to start everthing with "source" or
"harmonized" where appropriate. Make URLs more concise. Change
URL names to be shorter and more clear. Change view names to match
the new URL names. Change links in the navbar, home, and search
templates to compensate for the new changes. Change the URL names
in trait_browser.test_views to compensate. Change the LinkColumns
in the two django-tables2 objects to compensate.