Set up Selenium testing modules
Add a directory to hold selenium tests. Change the name of the test
module from functional_test.py to selenium_tests.py. Set up a
superclass for selenium testing. Set up two minimal selenium test
cases for the home page and the admin login page. Add the chromedriver
executable file for being able to run selenium tests with Chrome as
the driver. Add the selenium-server-standalone-2.53.0.jar file to
enable use of the Safari driver (The python bindings for the Safari
driver are not yet written for the selenium python package, so the
Safari driver uses the standalone server from this jar file.)