Add fake data to the test db for selenium test class and fill out HomeTest and AdminTest
Add code to SeleniumTestCase's setUp method to add lots of fake data
to the test database. Add creation of a superuser in the test database
to SeleniumTestCase's setUp and remove the overridden setUp method in
the AdminTest class.

Add code to HomeTest.test_home to click on each of the source phenotype
dropdown menu items in turn and navigate back to the home page. Add code
to AdminTest.test_admin to click on each of the model admin links in
turn and then navigate back.