Add check_table_view method to SeleniumTestCase, including check on expected_rows
Add a method check_table_view() to SeleniumTestCase class. This method
implements tests for pages that have django-tables2 pages. So far the
method checks that a table is on the page and, with the optional
argument expected_rows, checks that the total number of rows in the
table (across multiple pages) is correct.

Add a TablePageTestCase class to implement tests for the website
pages that have a table and should be tested with the check_table_view
function. So far only one test is implemented, as a test of the
check_table_view function.