WIP from when I was working on this in March - unknown purpose
Commit some changes to the test_selenium_development.py module, the
purpose of which I don't recall. I was in the middle of working on this
in March, but stopped before finishing the goal. I don't remember why
I implemented these changes now and I don't want to lose them. I'll undo
them in the next commit.