Add local.py settings for password email reset to work
This requires running the python internal mail server, which is good for
debug mode but not for production; that will be configured later. Run the
python mail server like so:
python -m smtpd -n -c DebuggingServer localhost:1025
in the background before running ./manage.py runserver. Then the link to
the password reset page will be printed out (replace topmedphenotypes.org
with localhost).

see:
http://garmoncheg.blogspot.com/2012/07/django-resetting-passwords-with.html