fix tests for @login_required decorator in views
Most of the test cases for the views needed to be updated to use a logged-in
user, or else we get a redirect to the login page instead of the actual view
to be tested. Here, add a TestCase class called ViewsAutoLoginTestCase that
creates a user and logs them in, and then make the view test cases subclasses
of that class. One of the TestCase classes in the test_views script
(SearchTestCase) doesn't need to be a subclass of ViewsAutoLoginTestCase,
because it is a helper function for a view and doesn't take an http request
object as an argument.