Fix errors in admin.py.
Move SourceDatasetUniqueKeys to the end of all the admin classes, to
match the new order in models.py.
Change the model name in each of the readonly_fields specifications,
which had been left as "Study" for most due to copy/pasting.
Remove most foreign key fields in list_display, list_filter, and
search_fields due to errors.