Make type fixing functions more concise
Add _fix_row helper function and replace all uses of the individual
fixer functions with _fix_row. This removes an issue where one of the
model's populate function was producing a warning about non-timezone-aware
datetimes being passed to it (because the _fix_timezone function had been
inadvertently not applied to it).