Implement source_dataset_subcohorts from snuffles as a ManyToManyField of SourceDataset model; remove SourceDatasetUniqueKeys
source_dataset_unique_keys and source_dataset_subcohorts from snuffles are actually
"junction tables", which can be implemented more straigtforwardly as ManyToManyFields
in Django models. Add subcohorts as a ManyToManyField of SourceDatasets to address this.
Could have added unique_keys as a ManyToManyField of SourceDatasets, but this revealed
that source_dataset_unique_keys might be better implemented as is_unique_key and
is_visit_column fields of SourceTrait objects, so put this on hold until I can talk about
it with Adrienne.