Move TOPMed logo inside of one jumbotron; Modify homepage warning
Put the TOPmed logo on the homepage into the same jumbotron as the
homepage title and description. Putting it after the text fixed spacing
issues that were preventing this before.

Change the warning message on the homepage so that it doesn't call the
site "beta" and doesn't mention that URLs might change.