Add mysqlclient back to pip requirements file
Because mysqlclient couldn't be installed on Leslie's desktop at the
time that it was added to the requirements, it was later removed from
the requirements list when another package was added. Add it back to
the requirements. Leslie has now been able to install mysqlclient
(by first installing the mysql package in macports), so this issue
should not recur.