Change import_db.py integration test to use 'devel' as source_db
Add the option which_db='test' for test_import_db.IntegrationTest.
test_call_command_to_import_whole_db(). This test previously used
the tomped_test_emeryl db, but with the new db structure on suwa that
is outdated.

Later, this will be changed to use the devel db and first load a
pre-specified test dataset. But for now the test db is the same
for everyone and already has data in it. Unfortunately, it does
not have any subcohort data, so the tests for loading subcohort
data currently fail.