Add restart_fix_wsgi.py that can be used to trigger an Apache server restart
Add this wsgi file that can be used to restart Apache. I used this recently
to force Apache to restart, and it worked well. Sometimes if you try
running a set of code that has an error, if that error occurs in the
wsgi in particular, it can prevent the Django code from working
properly later, even if you fix all of the problems with it.