Fix DEFINER lines in test source_db mysqldump files for portability
mysqldump files generated on another system have a problematic line in
the view creation statements because they state DEFINER=username@connecting_host.
Because the connecting host is different on another system, this is not
portable. To fix this, change these lines to read DEFINER=CURRENT_USER,
which will work on other systems.