Import component traits and adjust models for component traits linked to htrait sets
Currently, component traits are linked to harmonized_trait_sets rather than
to harmonized_traits (in snuffles). This will be changed later, but in
order to handle the component traits now, make a few changes that
can be easily undone later.

Also add the component source traits to the harmonized trait detail view.