Add user to Recipe models from form, and use one generic new recipe view
Make one view, new_recipe, to repace new_unit_recipe and new_harmonization_recipe,
with an argument to specify which kind of recipe to use.
Add form processing to the view, which includes saving the input data,
saving the m2m fields, and saving the value of the logged in user
as the creator and modifier.
Remove blank=True from the user related fields in the Recipe models.
Update the views specified in the urls for the recipe views.