Filter choices for units in harm. recipe form to those created by current user
Alter HarmonizationRecipeForm in __init__ so that the units field
is restricted to choices from the logged in user. Add a test to
ensure that choices from other users are rejected by the form. This
exposed an error in the testing method I had been using. All of the
tests of valid forms broke because I was adding the user attribute
to the form after instantiating it with input. So when __init__ was
called and the unit choices filtered to those by the current user,
user was set to none and no choices were valid. Add the arg
user=user to the form instantiation, which fixes this error in testing.