Make name fields shorter for both recipe models, to address MySQL issues
Change max_length of name fields for UnitRecipe and HarmonizationRecipe
to be 255 characters (previously 1000). Replace the old initial migration
file for the recipes app (creating a second migration doesn't cut it,
since the initial migration would still fail). This is to prevent a MySQL
error during migration.

When running the 0001_inital migration for recipes on the staging server,
I encountered this error: "Specified key was too long; max key length is
767 bytes". Googling revealed that this is related to the length of the
fields used in creating a unique together constraint. There is a limit
on the length of these fields because an index is created to maintain
the uniqueness, and there are constraints on the length of indexes in
InnoDB tables. The actual character length varies depending on the
character set being used. I wasn't sure exactly what the limit would be
for the character set we're using (latin1, I think). But trying 255 worked,
so I went with that.