Fix core.admin.UserAdmin so that it sends an email as needed
Change template name for subject template used by PasswordResetForm
in the UserAdmin. The UserAdmin was taken directly from a file
downloaded from the authtools documentation. The file was wrong, while
the code on the documentation website was correct. Fix the if
statement to match the corrected version on the website, which
results in password reset emails being sent appropriately.