Add site context processor and use it in the base settings
Add phenotype_inventory.context_processors, which contains the site()
processor. site() adds the site variable to every template used in the
entire project, which allows dynamic setting of the site name in the
navbar template. Use this site variable to set the site name in the
navbar template. Turn on the use of this context processor in the
base settings file.