Add test_email_sending.py and DEVELOPMENT custom setting
Add DEVELOPMENT to the base settings file. Use this to do things
that only need to be done on the staging and production servers, or
vice versa. Add test_email_sending.py, which contains a test that
will make sure an email is successfully sent to a temporary email
address, but this test is only run on the non-development systems.