Move login required test code to core.utils and reuse it in test_views
Since code to test login required of views was repeated in both recipes.test_views
and trait_browser.test_views, move it instead to core.utils and make the
class LoginRequiredTestCase there.

Move helper functions out of the LoginRequiredTestCase class so that
they could be used in other places as well.

Rewrite TraitBrowserLoginRequiredTestCase and RecipesLoginRequiredTestCase
to use the new base test case in utils.