Add SourceDBTimeStampedModel and compensate in factories
In trait_browser.models add the SourceDBTimeStampedModel abstract
base class for adding both the Django created and modified timestamps
as well as the i_date_added and i_date_changed timestamps from the
source db. Then make all of the trait_browser models subclasses of
SourceDBTimeStampedModel.

To compensate for this, all of the trait_browser factories need to
have the source db timestamp fields added to them as well, with
reasonable fake data. Make SourceDBTimeStampMixin to add these
fields to all of the factories. This is the change that required
updated versions of factory-boy and Faker, because the datetime
Faker providers didn't work quite right with timezone support in the
old version.