Add i_date_added and i_date_changed fields to importing in import_db.py
In import_db.py, add date_changed and date_added to the list of fields
to import (the source_field_names argument to the _make_args_mapping()
method, which is called within each of the _make_[object_name]_args()
methods). Also add these two fields to necessary places in
test_import_db.py so that tests are run on appropriate fields.