Use new _make_table_query helper to make the update query
In _make_query_for_rows_to_update, change the function to use the new
helper function _make_table_query. This is a little more DRY. Also
change the behavior in the else statement so that an empty result set
is returned if there are no previously-imported table rows to update.