Fix tests of update methods to check modified datetime properly
Previously, the tests in UpdateModelsTestCase were checking that
modified > created, but this test will ALWAYS past because the modified
date is always _just_ after the created date. Fix this by saving a datetime,
t1, that is before the udpate is made in the source db. Then check that
modified is greater than t1. Tests are still passing.