Make a mixin to open and close connections to the source db
Add OpenCloseDBMixin to test_import_db. This mixin adds setUp and tearDown
methods to TestCase classes. setUp opens a connection to the source db and
a cursor. tearDown closes the db connection and the cursor. This mixin is
now used by both BaseTestDataTestCase and VisitTestDataTestCase instead of
repeating the code across TestCase classes.