Add child_related_name arg to _make_m2m_link() function
_make_m2m_link() was previously using some complicated code to find which
of the fields on parent_model was the field related to child_model. However,
this method breaks down when there are multiple M2M fields related to the
same model, which there will be (component_age_traits and component_source_traits
and component_visit_traits). Also the code was pretty ugly and unreadable.
So use the child_related_name argument instead and require that the
field name be given to the function.