Test removal of source_dataset-subcohort link in IntegrationTest
In IntegrationTest.test_handle_with_updated_data(), add code to test
the removal of a previously existing link between a subcohort and a
source dataset, in conjunction with the addition of a new link for a
different source dataset and subcohort pair. Also change the way that
the subcohorts are picked out for this test function, and for the tests
in the UpdateModelsTestCase.