Reduce args in helper functions using **kwargs
In the helper functions for **kwargs, there was a lot of repetition of
function arguments. e.g. there might be three levels of helper function,
each asking for the same set of 10 arguments. In order to reduce the
repetition of arguments, make all of the functions accept **kwargs. Then
have each function require only those arguments that are used directly
within that function. For arguments that are just passed to another function,
use **kwargs. This drastically reduced the number of arguments.