Have update/import functions take a connection object rather than db type
_import_source_tables, _import_harmonized_tables, _update_source_tables,
and _update_harmonized_tables used to take which_db as an argument, which
was one of 'devel', 'test', or 'production'. But that prevents a single lock
function from locking the db to others' writing during the call to handle(),
so change these four functions to take the source_db argument instead. Then
call _get_source_db only once in handle() and pass the same source_db
connection object to each of the four import/update functions.