Prevent use if --import_only and --update_only outside of tests
If import or update functions were run separately on a production db,
it would cause missed data due to the way that latest_date is determined
from the existing Django db objects. But the --update_only and --import_only
options make it much easier to test the management command functions
without extra output from logging. So implement this somewhat-hacky way of detecting
that the command is being run in a test, and raising an error if --import_only
or --update_only is used outside of testing.