Keep null source db values if source db field is a date
Change import_db's _fix_null() so that fields with the word "date" in them
do not have their null values (None in python) replaced with empty strings.
The empty string works with the other field types for Django, but the Django
DateTimeField needs to take None for null values (rather than an empty string).
This is necessary to add the i_date_visit_reviewed field to the SourceDataset
model, since this is the first date field imported into PIE which can be
expected to be null often.