Check for correct imported pks in IntegrationTest
In IntegrationTest.test_handle_with_visit_data, assert statements used to
test that the number of imported objects matched the number of rows in the
source tables. Change that so that now the assert statements check that the
imported primary keys (pks) match the pks in the source db tables.