Add check_imported_values_match; use it in new integration test of updates
Add reusable test method that checks whether the imported field values for
a given model match the values found in the source db table they were taken from.
Use this method to make the test_values_match_after_all_updates integration test.
So far, this test only checks on the values in the regular models, and does not
yet check the values for the M2M relationships.