Improve _fix_null() by using data type
Several table fields in the source db are allowed to be null, especially those
that are designed to be filled with visit tracker data. Null values produce a
python None when pulling data from the source db in import_db with MySQL connector.
But Django stores string null values as empty strings (not None), so the _fix_null()
function is intended to convert imported None's into ""s. But this really only makes
sense for the string fields. For other nullable fields, especially datetime and boolean
fields, the None value is needed to get the proper NULL storage in the Django db.

Instead of having _fix_null convert all None's to empty strings, only have it convert
None's from string-link fields (of the types in STRING_TYPES). This requires adding
a field_types argument to _fix_null, which is also used by _fix_row. So also add the
use of field_types every time _fix_null or _fix_row is called, in import_db and
in its tests. And add a simulated field_types dictionary in db_factory's fake_row_dict().