Switch to using HarmonizedTraitSet detail view for harm. traits
Use HarmonizedTraitSetDetail view instead of HarmonizedTraitDetail. Change
to this view in the urls, set a get_absolute_url method for HarmonizedTraitSet,
and switch to using this for the link column in HarmonizedTraitTable. Also make
a basic version of the detail view template.

I was trying to modify the HarmonizedTraitDetail template to make it show
all of the harmonization unit and component traits information, but this
brought up some complications, and I decided it would be best to just make
one detail page for each harmonized trait set and show all of the information
there.