Change the way that modified times are compared in test_import_db
Used to save a time, t1, and then check that a model's modified time was
later than this. But testing this on the staging server, things go much
faster and the two times are too close to compare. So instead save the old
modified time and make that comparison instead.