Allow full connection to production db, with locking only arg
When I had to add the source db locking functions, I had forgotten that
connecting to the production db with full permissions was specifically
disallowed in the get_db function. Now allow this connection, as long as
the "just_locking" argument is set, so that it doesn't happen by accident.