Fix staging server detail URL rendering on HarmonizedTraitSet detail page
Detail URLs for SourceTrait, HarmonizedTrait, and Study objects on the
HarmonizedTraitSetDetail page are stored in the HarmonizedTraitSet.component_html_detail
model field. Using the get_absolute_url method (counterintuitively) gives
their relative URL, so these URLs were broken on the staging server
where the site is at /pheno. Add a custom template tag to render a template
variable as template code, so that the {% url %} tag can be used in these
html strings.