Fix INTERNAL_IPS setting for debug toolbar on staging server
In settings.staging, set the INTERNAL_IPS to a tuple containing the IP
address of gcc-pc-004, the machine serving the staging site. The previous
settings, using an always-True show_toolbar() function, did correctly display
the toolbar on the staging server, but it caused all view tests to fail
with this error:
django.core.urlresolvers.NoReverseMatch: 'djdt' is not a registered namespace
I'm not quite sure why the error was occurring, but using the INTERNAL_IPS
setting and the default method for determining whether to show the toolbar
fixed the issue.