Fix selenium test that broke after changing sort order of source traits
Fix selenium_tests.test_selenium_development.TablePageTestCase.test_source_all_table,
which was broken because it depended on the sort order of the table of source traits.
Change the way the test works so that it just finds the first listed trait and checks
that detail page link, instead of getting the first trait from the db.