Fix h.t. encoded value importing error
Previously, importing and upating of harmonized trait encoded values was
erroneously using the harmonized trait id as the primary key of the
HarmonizedTraitEncodedValue object. Change the specification of the
source_pk argument in the calls to _update_existing_data and _import_new_data
for the harmonized encoded values. Fixes a failing test.