Don't change the underlying db column name of the global study id field
In trying to run migration 6 (a monstrous migration) on the staging server, I found
that it failed with a MySQL error about trying to change a field of the global study
table. I eventually figured out that changing the GlobalStudy.study_id field to
GlobalStudy.i_id was causing a problem, because other models use that field as a foreign
key. To prevent this issue, I kept the change in field name, but used the db_column
argument to the field declaration to tell Django to use the existing study_id column in
the underlying database. So this commit includes that change and the updated migration
file to accommodate that change.