Update chromedriver to fix failing Selenium tests
Selenium tests using the SeleniumTestCase.check_table_view() function were
failing with this error message:

selenium.common.exceptions.WebDriverException: Message: unknown error: Element is not clickable at point

It appears there was a bug in the chromedriver with my current version of
Chrome that prevented Selenium from being able to click on a button that was
not in the initial view of the webpage. So the check_table_view function
found the next button at the bottom of a table, but couldn't click on it
because it wasn't in the view field of the webpage to begin with.
Update the version of chromedriver to fix these failing tests.