Add tags models to the the admin site, and update Selenium admin site tests
Set up ModelAdmin classes for Tag and TaggedTrait models. Register those models
with the admin site. Went to add these to Selenium testing of the admin site and
found that no tests had been added for newer trait_browser or recipes models. So
split up the tests in AdminTestCase into separate tests for each of the apps.