Add a TaggedTrait create view, with related changes
Add the TaggedTraitForm class to tags.forms. Add the TaggedTraitCreate view
and its url and template. Add tests for the form and for the view.

Remove the default=False argument in the Tag.recommended field. Because it's
a BooleanField, the defaults and blank behavior don't work exactly as I'd
previously expected. You don't want blank=True here, because you do need the
form to specify whether recommended is true or false. But you also don't want
to have default=False, because that decision should be handled at object
declaration, not at the database level. Regenerate the initial migration for
tags to account for this change.