Customize text on taggedtrait_form for each form
Add title attribute to each of the forms that is displayed using the
taggedtrait_form template. Display that title in the template itself,
and display either the tag or the trait depending on which (if any) is
found in the context. Add the tag or the trait to the context of the
view for the two cases where it's needed. Check for the context in the
view tests as appropriate.