Fix login_required testing for pk-specific autocomplete views
The TraitBrowserLoginRequiredTestCase was failing because the new URL for
autocompleting source traits from a specific study doesn't work when the
study doesn't exist yet. So add a call to build_test_db() before running
the assert statement in this test. Also add a URL-specific error message
to the call to assertRedirects in the LoginRequiredTestCase.