Add login required test for tags views, and try to fix pk issue
Add TagsLoginRequiredTestCase. But now login required tests are failing for recipes,
tags, and trait browser, due to issues with pks in URLs and auto-incrementing in
the test db. Briefly, LoginRequiredTestCase just puts a 1 into each URL that needs
a pk, but because the test db is not destroyed and recreated after each test (just
emptied after each test), the auto-increment value doesn't reset either. This may
be related to the recent switch from SQLite3 to MariaDB in development. There is a
(somewhat complicated) way to get from a URL to the view class that it's related to
and from there to the model you need to use to get a valid pk from, but it will
only work with Django 1.9+. So I added that code, but commented out, and these
tests will just keep failing until I do the upgrade to 1.11.