Reorganize URLs in tags and make app URLs plural in root url includes
In the root urls.py, make all of the namespaces and URLs for the custom apps
plural. E.g. change recipe to recipes. Also reorganize the order and comments
in the root urls.py file. Reorganize the tags urls.py file to reduce repetition
of "tag" and "tagging". It's still incomplete as I now need to write several
new forms and views.