Add tagging restrictions and tests of them for the trait_browser tagging views
Add UserPassesTestMixin and a test_func that ensures the trait being tagged is in the user's
taggable_studies list to the SourceTraitTagging view. Add the TaggableStudiesRequiredMixin and
GroupRequiredMixin with required 'phenotype_taggers' to the
TaggableStudyFilteredSourceTraitPHVAutocomplete view. Add tests of these permissions and other
restrictions to the tests of these two views in trait_browser.test_views.