Tag phenotype button only displays for appropriate users
In the source trait detail page, only display the "tag phenotypes" button if the
user is a phenotype tagger with the appropriate taggable_studies or if the user is
a DCC staff member. To do this, add a variable user_is_study_tagger to the detail
view's context. Also update tests for the source trait detail view.