Set up recommended_traits field in ManyTaggedTraits forms and views
Remove the recommended field and add the recommended_traits field to the
ManyTaggedTraitsForm and ManyTaggedTraitsByTagForm. Add a clean method to
both forms to make the form require at least one trait in one or the other
of the fields. Modify existing form tests and view tests for these changes.
Add tests of the form cleaning that requires at least one (but not both)
traits.