Set up tagged trait list views by study
Add table classes for displaying tallies of tags and traits by study and
for displaying tagged traits. Add views	for the	study tag summary page and
a study-specific list of tagged	traits.	Add tests of the new tables and
views. Add urls for the new views. Add a template for the new tagged trait
list view.